moved few functions to new js file
